bot/requirements.txt
# Bot Requirements
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pillow>=10.0.0
stripe==7.8.0